Various bug fixes:  - RTL8019 overflow problem fixed  - ARP bugs fixed  - HTTP server fixed so it keeps sockets listening  - Less #include dependencies  - TCP sochandle vars are now all signed

+++ system.c
@@ -115,7 +115,7 @@ Desc:			Calculates the length of given string
 
 INT16 strlen (UINT8* buf, UINT16 len)
 {
-	UINT8 i;
+	UINT16 i;
 	
 	for(i=0; i<len; i++) {
 		if(*buf == '\0')
